renames a few scopes
