+++ internal/handlers/slashEvents/commands/profile/me.go
@@ -5,11 +5,11 @@ import (
 	"log"
 	"time"
 
-	"github.com/RazvanBerbece/Aztebot/internal/api/member"
-	rolesService "github.com/RazvanBerbece/Aztebot/internal/api/roles"
 	dataModels "github.com/RazvanBerbece/Aztebot/internal/data/models"
 	globalConfiguration "github.com/RazvanBerbece/Aztebot/internal/globals/configuration"
 	globalRepositories "github.com/RazvanBerbece/Aztebot/internal/globals/repositories"
+	"github.com/RazvanBerbece/Aztebot/internal/services/member"
+	rolesService "github.com/RazvanBerbece/Aztebot/internal/services/roles"
 	"github.com/RazvanBerbece/Aztebot/pkg/shared/embed"
 	"github.com/RazvanBerbece/Aztebot/pkg/shared/utils"
 	"github.com/bwmarrin/discordgo"
