+++ internal/handlers/slashEvents/commands/staff/jail/unjail.go
@@ -3,9 +3,9 @@ package jailSlashHandlers
 import (
 	"fmt"
 
-	"github.com/RazvanBerbece/Aztebot/internal/api/member"
 	dataModels "github.com/RazvanBerbece/Aztebot/internal/data/models"
 	globalConfiguration "github.com/RazvanBerbece/Aztebot/internal/globals/configuration"
+	"github.com/RazvanBerbece/Aztebot/internal/services/member"
 	"github.com/RazvanBerbece/Aztebot/pkg/shared/embed"
 	"github.com/RazvanBerbece/Aztebot/pkg/shared/utils"
 	"github.com/bwmarrin/discordgo"
