+++ internal/handlers/slashEvents/commands/staff/timeout/timeoutAppeal.go
@@ -3,8 +3,8 @@ package timeoutSlashHandlers
 import (
 	"fmt"
 
-	"github.com/RazvanBerbece/Aztebot/internal/api/member"
 	globalConfiguration "github.com/RazvanBerbece/Aztebot/internal/globals/configuration"
+	"github.com/RazvanBerbece/Aztebot/internal/services/member"
 	"github.com/RazvanBerbece/Aztebot/pkg/shared/embed"
 	"github.com/RazvanBerbece/Aztebot/pkg/shared/utils"
 	"github.com/bwmarrin/discordgo"
