+++ internal/services/tasks/startup/publishDefaultEmbeds.go
